requirements.txt
fastapi==0.115.4
sqlmodel==0.0.22
PyJWT==2.9.0
passlib==1.7.4
pydantic-settings==2.6.1
